Code refractor done
