+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/CommonUi.kt
@@ -60,91 +60,6 @@ val intensityBorder = listOf(
     Color(0xFFF44336)    // 6 - Red Border
 )
 
-fun labelColor(intensity: Int) =
-    if (intensity >= 2) Color(0xFF111111) else Color(0xFFFFFFFF)
-
-val statusLabel = listOf("", "LOW", "MED", "HIGH", "ON", "ON", "ON")
-
-@Composable
-fun IntensityWheelPicker(
-    intensity: Int,
-    onIntensityChange: (Int) -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    isRed: Boolean = false
-) {
-    val states = if (isRed) listOf(0, 3) else listOf(0, 1, 2, 3)
-    val state = rememberLazyListState(initialFirstVisibleItemIndex = states.indexOf(intensity).coerceAtLeast(0))
-
-    // Sync external state changes to picker
-    LaunchedEffect(intensity) {
-        val targetIdx = states.indexOf(intensity)
-        if (targetIdx >= 0 && state.firstVisibleItemIndex != targetIdx) {
-            state.animateScrollToItem(targetIdx)
-        }
-    }
-
-    // Sync picker changes to external state
-    LaunchedEffect(state.isScrollInProgress) {
-        if (!state.isScrollInProgress) {
-            val centerIdx = state.firstVisibleItemIndex
-            if (centerIdx in states.indices) {
-                onIntensityChange(states[centerIdx])
-            }
-        }
-    }
-
-    Box(
-        modifier = modifier
-            .size(70.dp)
-            .clip(RoundedCornerShape(12.dp))
-            .background(Color(0xFF111111))
-            .border(1.dp, Color(0xFF222222), RoundedCornerShape(12.dp)),
-        contentAlignment = Alignment.Center
-    ) {
-        LazyRow(
-            state = state,
-            flingBehavior = rememberSnapFlingBehavior(lazyListState = state),
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(horizontal = 15.dp),
-            userScrollEnabled = enabled
-        ) {
-            items(states.size) { idx ->
-                val level = states[idx]
-                val isSelected = level == intensity
-                
-                Box(
-                    modifier = Modifier
-                        .fillMaxHeight()
-                        .width(40.dp)
-                        .padding(4.dp),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .size(if (isSelected) 24.dp else 12.dp)
-                            .clip(CircleShape)
-                            .background(intensityColor[if (isRed && level > 0) 6 else level])
-                            .border(
-                                width = 1.dp,
-                                color = if (isSelected) Color.White else Color.Transparent,
-                                shape = CircleShape
-                            )
-                    )
-                }
-            }
-        }
-        
-        // Center indicator overlay
-        Box(
-            Modifier
-                .fillMaxHeight()
-                .width(40.dp)
-                .border(1.dp, Color.White.copy(alpha = 0.1f), RoundedCornerShape(4.dp))
-        )
-    }
-}
-
 @Composable
 fun SectionLabel(text: String) {
     Text(text, color = Color(0xFF666666), fontSize = 11.sp, modifier = Modifier.fillMaxWidth())
