+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/composer/components/common/StepPreviewBox.kt
@@ -1,0 +1,155 @@
+package com.smaarig.glyphbarcomposer.ui.composer.components.common
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.detectDragGestures
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.DragHandle
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.smaarig.glyphbarcomposer.model.GlyphSequence
+import com.smaarig.glyphbarcomposer.ui.getChannelForIndex
+import com.smaarig.glyphbarcomposer.ui.intensityBorder
+import com.smaarig.glyphbarcomposer.ui.intensityColor
+
+@Composable
+fun StepPreviewBox(
+    step: GlyphSequence,
+    onDelete: () -> Unit,
+    enabled: Boolean,
+    onDragStart: () -> Unit = {},
+    onDragEnd: () -> Unit = {},
+    onDragCancel: () -> Unit = {},
+    onDrag: (change: androidx.compose.ui.input.pointer.PointerInputChange, dragAmount: androidx.compose.ui.geometry.Offset) -> Unit = { _, _ -> },
+    modifier: Modifier = Modifier
+) {
+    var showMenu by remember { mutableStateOf(false) }
+
+    Box(
+        modifier = modifier
+            .fillMaxWidth()
+            .height(64.dp)
+            .clip(RoundedCornerShape(16.dp))
+            .background(
+                brush = Brush.horizontalGradient(
+                    colors = listOf(Color(0xFF181818), Color(0xFF111111))
+                )
+            )
+            .border(1.dp, Color(0xFF2A2A2A), RoundedCornerShape(16.dp))
+            .clickable(enabled = enabled) { showMenu = !showMenu },
+        contentAlignment = Alignment.Center
+    ) {
+        Row(
+            modifier = Modifier
+                .padding(horizontal = 14.dp, vertical = 10.dp)
+                .fillMaxSize(),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(12.dp)
+            ) {
+                Icon(
+                    Icons.Default.DragHandle,
+                    contentDescription = "Drag to reorder",
+                    tint = Color(0xFF555555),
+                    modifier = Modifier
+                        .size(32.dp)
+                        .padding(4.dp)
+                        .pointerInput(Unit) {
+                            detectDragGestures(
+                                onDragStart = { onDragStart() },
+                                onDragEnd = { onDragEnd() },
+                                onDragCancel = { onDragCancel() },
+                                onDrag = { change, dragAmount -> onDrag(change, dragAmount) }
+                            )
+                        }
+                )
+
+                Column(verticalArrangement = Arrangement.Center) {
+                    Text(
+                        "STEP",
+                        fontSize = 8.sp,
+                        color = Color(0xFF666666),
+                        fontWeight = FontWeight.Black,
+                        letterSpacing = 0.5.sp
+                    )
+                    Text(
+                        "${step.durationMs.toInt()}ms",
+                        fontSize = 12.sp,
+                        color = Color.White,
+                        fontWeight = FontWeight.Bold
+                    )
+                }
+            }
+            Row(
+                horizontalArrangement = Arrangement.spacedBy(4.dp),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                repeat(7) { i ->
+                    val intensityVal = step.channelIntensities[getChannelForIndex(i)] ?: 0
+                    val finalIntensity = if (i == 6 && intensityVal > 0) 6 else intensityVal
+                    
+                    Box(
+                        modifier = Modifier
+                            .size(width = 6.dp, height = 24.dp)
+                            .clip(RoundedCornerShape(1.5.dp))
+                            .background(intensityColor[finalIntensity])
+                            .border(
+                                0.5.dp,
+                                intensityBorder[finalIntensity].copy(alpha = 0.5f),
+                                RoundedCornerShape(1.5.dp)
+                            )
+                    )
+                }
+            }
+        }
+
+        if (showMenu) {
+            Box(
+                Modifier
+                    .fillMaxSize()
+                    .background(Color.Black.copy(alpha = 0.85f))
+                    .clickable { showMenu = false },
+                contentAlignment = Alignment.Center
+            ) {
+                Row(
+                    horizontalArrangement = Arrangement.spacedBy(20.dp),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    IconButton(
+                        onClick = { onDelete(); showMenu = false },
+                        modifier = Modifier
+                            .size(36.dp)
+                            .background(Color(0xFF221111), CircleShape)
+                            .border(1.dp, Color(0xFF442222), CircleShape)
+                    ) {
+                        Icon(
+                            Icons.Default.Delete,
+                            null,
+                            tint = Color(0xFFFF5252),
+                            modifier = Modifier.size(18.dp)
+                        )
+                    }
+                }
+            }
+        }
+    }
+}
