+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/composer/components/v2/DraggableTimeline.kt
@@ -1,0 +1,196 @@
+package com.smaarig.glyphbarcomposer.ui.composer.components.v2
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.PlayArrow
+import androidx.compose.material.icons.filled.Save
+import androidx.compose.material.icons.filled.Stop
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.zIndex
+import com.smaarig.glyphbarcomposer.ui.composer.components.common.EmptyTimelinePlaceholder
+import com.smaarig.glyphbarcomposer.ui.composer.components.common.StepPreviewBox
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerUiState
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerViewModel
+import kotlin.math.roundToInt
+
+@Composable
+fun DraggableTimeline(
+    uiState: ComposerUiState,
+    viewModel: ComposerViewModel,
+    modifier: Modifier = Modifier
+) {
+    var draggingIndex by remember { mutableStateOf<Int?>(null) }
+    var dragOffsetY by remember { mutableFloatStateOf(0f) }
+    val listState = rememberLazyListState()
+    var showSaveDialog by remember { mutableStateOf(false) }
+    var fileName by remember { mutableStateOf("") }
+
+    if (showSaveDialog) {
+        AlertDialog(
+            onDismissRequest = { showSaveDialog = false },
+            title = { Text("Save Sequence", color = Color.White, fontWeight = FontWeight.Black) },
+            text = {
+                TextField(
+                    value = fileName,
+                    onValueChange = { fileName = it },
+                    placeholder = { Text("Sequence Name", color = Color.Gray) },
+                    singleLine = true,
+                    colors = TextFieldDefaults.colors(
+                        focusedContainerColor = Color(0xFF1A1A1A),
+                        unfocusedContainerColor = Color(0xFF1A1A1A),
+                        focusedTextColor = Color.White,
+                        unfocusedTextColor = Color.White,
+                        cursorColor = Color.White
+                    )
+                )
+            },
+            confirmButton = {
+                TextButton(onClick = {
+                    if (fileName.isNotBlank()) {
+                        viewModel.savePlaylist(fileName)
+                        showSaveDialog = false
+                        fileName = ""
+                    }
+                }) {
+                    Text("SAVE", color = Color(0xFF00C853), fontWeight = FontWeight.Black)
+                }
+            },
+            dismissButton = {
+                TextButton(onClick = { showSaveDialog = false }) {
+                    Text("CANCEL", color = Color.Gray)
+                }
+            },
+            containerColor = Color(0xFF111111),
+            shape = RoundedCornerShape(28.dp)
+        )
+    }
+
+    val itemHeightPx = 183f
+
+    Column(
+        modifier = modifier
+            .clip(RoundedCornerShape(24.dp))
+            .background(Color(0xFF080808))
+            .border(1.dp, Color(0xFF1A1A1A), RoundedCornerShape(24.dp))
+            .padding(14.dp),
+        verticalArrangement = Arrangement.spacedBy(10.dp)
+    ) {
+        Text(
+            "SEQUENCE",
+            color = Color(0xFF666666),
+            fontSize = 10.sp,
+            fontWeight = FontWeight.Black,
+            letterSpacing = 1.sp
+        )
+
+        Box(modifier = Modifier.weight(1f)) {
+            if (uiState.currentSequenceSteps.isEmpty()) {
+                EmptyTimelinePlaceholder()
+            } else {
+                LazyColumn(
+                    state = listState,
+                    modifier = Modifier.fillMaxSize(),
+                    verticalArrangement = Arrangement.spacedBy(6.dp)
+                ) {
+                    itemsIndexed(uiState.currentSequenceSteps) { index, step ->
+                        val isDragging = draggingIndex == index
+
+                        Box(
+                            modifier = Modifier
+                                .zIndex(if (isDragging) 1f else 0f)
+                                .graphicsLayer {
+                                    translationY = if (isDragging) dragOffsetY else 0f
+                                    alpha = if (isDragging) 0.78f else 1f
+                                    scaleX = if (isDragging) 1.03f else 1f
+                                    scaleY = if (isDragging) 1.03f else 1f
+                                }
+                        ) {
+                            StepPreviewBox(
+                            step = step,
+                            onDelete = { viewModel.removeStep(index) },
+                            enabled = !uiState.isPlaying,
+                            onDragStart = {
+                                if (!uiState.isPlaying) draggingIndex = index
+                            },
+                            onDragEnd = {
+                                val src = draggingIndex
+                                if (src != null) {
+                                    val steps = (dragOffsetY / itemHeightPx).roundToInt()
+                                    val dst = (src + steps)
+                                        .coerceIn(0, uiState.currentSequenceSteps.size - 1)
+                                    if (dst != src) viewModel.reorderSteps(src, dst)
+                                }
+                                draggingIndex = null
+                                dragOffsetY = 0f
+                            },
+                            onDragCancel = {
+                                draggingIndex = null
+                                dragOffsetY = 0f
+                            },
+                            onDrag = { change, amount ->
+                                change.consume()
+                                if (!uiState.isPlaying) dragOffsetY += amount.y
+                            }
+                        )
+                        }
+                    }
+                }
+            }
+        }
+
+        if (uiState.currentSequenceSteps.isNotEmpty() && draggingIndex == null) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(40.dp),
+                horizontalArrangement = Arrangement.spacedBy(8.dp)
+            ) {
+                IconButton(
+                    onClick = {
+                        if (uiState.isPlaying) viewModel.stopPlayback()
+                        else viewModel.startPlayback(uiState.currentSequenceSteps)
+                    },
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxHeight()
+                        .background(
+                            if (uiState.isPlaying) Color(0xFF00E676) else Color.White,
+                            RoundedCornerShape(10.dp)
+                        )
+                ) {
+                    Icon(
+                        if (uiState.isPlaying) Icons.Default.Stop else Icons.Default.PlayArrow,
+                        null,
+                        tint = Color.Black,
+                        modifier = Modifier.size(20.dp)
+                    )
+                }
+                IconButton(
+                    onClick = { showSaveDialog = true },
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxHeight()
+                        .background(Color(0xFF1A1A1A), RoundedCornerShape(10.dp))
+                ) {
+                    Icon(Icons.Default.Save, null, tint = Color.White, modifier = Modifier.size(20.dp))
+                }
+            }
+            Spacer(Modifier.height(12.dp))
+        }
+    }
+}
