+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/composer/components/v2/LandscapeControlsRow.kt
@@ -1,0 +1,76 @@
+package com.smaarig.glyphbarcomposer.ui.composer.components.v2
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerUiState
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerViewModel
+
+@Composable
+fun LandscapeControlsRow(
+    uiState: ComposerUiState,
+    viewModel: ComposerViewModel,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(20.dp))
+            .background(Color(0xFF111111))
+            .border(1.dp, Color(0xFF222222), RoundedCornerShape(20.dp))
+            .padding(horizontal = 16.dp, vertical = 8.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(24.dp)
+    ) {
+        Column(modifier = Modifier.weight(1f)) {
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text("DURATION", color = Color.Gray, fontSize = 9.sp, fontWeight = FontWeight.Bold)
+                Text("${uiState.durationMs.toInt()}ms", color = Color.White, fontSize = 9.sp, fontWeight = FontWeight.Black)
+            }
+            Slider(
+                value = uiState.durationMs,
+                onValueChange = viewModel::onDurationChange,
+                valueRange = 100f..2000f,
+                steps = 18,
+                colors = SliderDefaults.colors(
+                    thumbColor = Color.White,
+                    activeTrackColor = Color.White,
+                    inactiveTrackColor = Color(0xFF222222)
+                )
+            )
+        }
+
+        Button(
+            onClick = viewModel::addStep,
+            modifier = Modifier
+                .height(44.dp)
+                .width(120.dp),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = Color.White,
+                contentColor = Color.Black
+            ),
+            shape = RoundedCornerShape(10.dp),
+            contentPadding = PaddingValues(0.dp)
+        ) {
+            Icon(Icons.Default.Add, null, modifier = Modifier.size(16.dp))
+            Spacer(Modifier.width(4.dp))
+            Text("ADD STEP", fontWeight = FontWeight.Black, fontSize = 10.sp)
+        }
+    }
+}
