+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/composer/components/v2/LandscapeGlyphsRow.kt
@@ -1,0 +1,84 @@
+package com.smaarig.glyphbarcomposer.ui.composer.components.v2
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.smaarig.glyphbarcomposer.ui.composer.components.common.GlyphScrollPicker
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerUiState
+import com.smaarig.glyphbarcomposer.ui.viewmodel.ComposerViewModel
+import com.smaarig.glyphbarcomposer.ui.viewmodel.RedGlyphViewModel
+
+@Composable
+fun LandscapeGlyphsRow(
+    uiState: ComposerUiState,
+    viewModel: ComposerViewModel,
+    redViewModel: RedGlyphViewModel,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(20.dp))
+            .background(Color(0xFF111111))
+            .border(1.dp, Color(0xFF222222), RoundedCornerShape(20.dp))
+            .padding(12.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Text("GLYPHS", color = Color.Gray, fontSize = 9.sp, fontWeight = FontWeight.Bold)
+        Spacer(Modifier.height(12.dp))
+
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .horizontalScroll(rememberScrollState()),
+            horizontalArrangement = Arrangement.spacedBy(12.dp, Alignment.CenterHorizontally),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            repeat(7) { index ->
+                val isRed = index == 6
+                val isSelected = uiState.selectedChannelIndex == index
+                val intensity = uiState.glyphIntensities[index]
+
+                if (isRed) {
+                    Box(modifier = Modifier.width(1.dp).height(44.dp).background(Color(0xFF2A2A2A)))
+                }
+
+                Column(
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                    verticalArrangement = Arrangement.spacedBy(6.dp)
+                ) {
+                    Box(
+                        modifier = Modifier
+                            .size(4.dp)
+                            .clip(CircleShape)
+                            .background(if (isSelected) Color.White else Color.Transparent)
+                    )
+                    GlyphScrollPicker(
+                        intensity = intensity,
+                        onIntensityChange = { newVal ->
+                            viewModel.onIntensityChange(index, newVal)
+                            viewModel.setSelectedChannel(index)
+                            if (isRed) redViewModel.setRed(newVal > 0)
+                        },
+                        isRed = isRed,
+                        enabled = !uiState.isPlaying
+                    )
+                }
+            }
+        }
+    }
+}
