+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/library/components/EmptyStateView.kt
@@ -1,0 +1,24 @@
+package com.smaarig.glyphbarcomposer.ui.library.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun EmptyStateView(icon: ImageVector, title: String, description: String) {
+    Column(modifier = Modifier.fillMaxWidth().padding(vertical = 40.dp), horizontalAlignment = Alignment.CenterHorizontally, verticalArrangement = Arrangement.Center) {
+        Icon(icon, null, tint = Color(0xFF1A1A1A), modifier = Modifier.size(64.dp))
+        Spacer(Modifier.height(16.dp))
+        Text(title, color = Color.Gray, fontWeight = FontWeight.Black, fontSize = 16.sp)
+        Text(description, color = Color(0xFF333333), fontSize = 12.sp, fontWeight = FontWeight.Bold, textAlign = TextAlign.Center)
+    }
+}
