+++ app/src/main/java/com/smaarig/glyphbarcomposer/ui/patternlab/components/SaveMixDialog.kt
@@ -1,0 +1,47 @@
+package com.smaarig.glyphbarcomposer.ui.patternlab.components
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SaveMixDialog(
+    onSave: (String) -> Unit,
+    onDismiss: () -> Unit
+) {
+    var name by remember { mutableStateOf("") }
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text("Save Mix", color = Color.White, fontWeight = FontWeight.Black) },
+        text = {
+            TextField(
+                value = name,
+                onValueChange = { name = it },
+                placeholder = { Text("Enter mix name", color = Color.Gray) },
+                singleLine = true,
+                colors = TextFieldDefaults.colors(
+                    focusedContainerColor = Color(0xFF1A1A1A),
+                    unfocusedContainerColor = Color(0xFF1A1A1A),
+                    focusedTextColor = Color.White,
+                    unfocusedTextColor = Color.White
+                )
+            )
+        },
+        confirmButton = {
+            TextButton(
+                onClick = { onSave(name) },
+                enabled = name.isNotBlank()
+            ) {
+                Text("SAVE", color = if (name.isNotBlank()) Color(0xFF00C853) else Color.Gray, fontWeight = FontWeight.Black)
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) { Text("CANCEL", color = Color.Gray) }
+        },
+        containerColor = Color(0xFF111111),
+        shape = RoundedCornerShape(28.dp)
+    )
+}
